+++ NEWS.md
@@ -4,6 +4,10 @@ output: html_document
 ---
 
 
+* Renaming and aggregation functions now (by default) infer 
+  row and column notations.
+
+
 # matsbyname 0.5.0 (2022-04-01) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6407149.svg)](https://doi.org/10.5281/zenodo.6407149)
 
 * New format for documentation pages,